# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewPack API Documentation",
        "description": "Documentación completa de la API de NewPack",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/suppliers/{supplier}/bank-accounts": {
            "get": {
                "tags": [
                    "Supplier Bank Accounts"
                ],
                "summary": "Get all bank accounts for a supplier",
                "operationId": "d94ed499afc4f22520613360ca60af79",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank accounts list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupplierBankAccountResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Supplier Bank Accounts"
                ],
                "summary": "Create a bank account for a supplier",
                "operationId": "b5e2fb16034a7fc8916fc25e4c1171c7",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bank_name"
                                ],
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "example": "BBVA"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "0123456789",
                                        "nullable": true
                                    },
                                    "clabe": {
                                        "type": "string",
                                        "example": "012180001234567890",
                                        "nullable": true
                                    },
                                    "swift_code": {
                                        "type": "string",
                                        "example": "BBVAMXMMXXX",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "MXN"
                                    },
                                    "is_primary": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Cuenta principal",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bank account created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierBankAccountResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers/{supplier}/bank-accounts/{bankAccount}": {
            "put": {
                "tags": [
                    "Supplier Bank Accounts"
                ],
                "summary": "Update a bank account",
                "operationId": "8696bcb94699341cc888bada5fe5fe51",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bankAccount",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bank_name": {
                                        "type": "string",
                                        "example": "BBVA"
                                    },
                                    "account_number": {
                                        "type": "string",
                                        "example": "0123456789",
                                        "nullable": true
                                    },
                                    "clabe": {
                                        "type": "string",
                                        "example": "012180001234567890",
                                        "nullable": true
                                    },
                                    "swift_code": {
                                        "type": "string",
                                        "example": "BBVAMXMMXXX",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "MXN"
                                    },
                                    "is_primary": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ],
                                        "example": "active"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Cuenta actualizada",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bank account updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierBankAccountResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Supplier Bank Accounts"
                ],
                "summary": "Delete a bank account",
                "operationId": "b2615367be07f3276c12e57b6290f043",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bankAccount",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/suppliers/{supplier}/bank-accounts/{bankAccount}/set-primary": {
            "post": {
                "tags": [
                    "Supplier Bank Accounts"
                ],
                "summary": "Set a bank account as primary",
                "operationId": "030c5741051cc616c168b9b2895058fa",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bankAccount",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bank account set as primary",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplierBankAccountResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SupplierBankAccountResource": {
                "title": "SupplierBankAccount",
                "description": "Supplier bank account resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "supplierId": {
                        "type": "integer",
                        "example": 1
                    },
                    "bankName": {
                        "type": "string",
                        "example": "BBVA"
                    },
                    "accountNumber": {
                        "type": "string",
                        "example": "0123456789",
                        "nullable": true
                    },
                    "clabe": {
                        "type": "string",
                        "example": "012180001234567890",
                        "nullable": true
                    },
                    "swiftCode": {
                        "type": "string",
                        "example": "BBVAMXMMXXX",
                        "nullable": true
                    },
                    "currency": {
                        "type": "string",
                        "example": "MXN"
                    },
                    "isPrimary": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "example": "active"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Cuenta principal",
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}